added better listing page.
Changed files:

app/hire/listings/page.tsx
components/features/hire/listings/index.ts
components/features/hire/listings/listings-createjob-modal.tsx
components/features/hire/listings/listings-delete-modal.tsx
components/features/hire/listings/listings-details-panel.tsx
components/features/hire/listings/listings-job-panel.tsx
components/features/hire/listings/listings-search-bar.tsx
hooks/hire/listings/use-listings-business-logic.ts
