+++ components/features/hire/listings/index.ts
@@ -1,0 +1,8 @@
+//dru dont delete, this is for making it easier for us to actually use these components in the future.
+//best practice according to AI lol
+
+export { ListingsSearchBar } from './listings-search-bar';
+export { ListingsJobPanel } from './listings-job-panel';
+export { ListingsDetailsPanel } from './listings-details-panel';
+export { ListingsCreateJobModal } from './listings-createjob-modal';
+export { ListingsDeleteModal } from './listings-delete-modal';
